refactor: i18n 파일들 위치 이동
