+++ src/app/[locale]/layout.tsx
@@ -4,7 +4,7 @@ import { Noto_Sans, Noto_Sans_JP, Noto_Sans_SC } from "next/font/google";
 import { ReactNode } from "react";
 
 import { META } from "@/components/const";
-import i18nConfig from "@/i18nConfig";
+import { i18nConfig } from "@/i18n";
 
 export function generateStaticParams() {
   return i18nConfig.locales.map((locale) => ({ locale }));
