+++ src/components/common/CommonProvider.tsx
@@ -1,6 +1,6 @@
-import TranslationsProvider from "@/components/common/TranslationsProvider";
 import GlobalStyle from "@/globalStyles";
 import { OverlayProvider } from "@/hooks/useOverlay";
+import { TranslationsProvider } from "@/i18n";
 
 const CommonProvider = ({ children, TranslationsProviderProps }) => {
   return (
