+++ src/components/user/LanguageSelector.tsx
@@ -4,7 +4,7 @@ import { usePathname, useRouter } from "next/navigation";
 import { useTranslation } from "react-i18next";
 
 import Icons from "@/components/common/Icons";
-import i18nConfig from "@/i18nConfig";
+import { i18nConfig } from "@/i18n";
 
 const StyledLanguageSelectorBox = styled.div`
   display: flex;
