+++ src/i18n/index.js
@@ -1,0 +1,3 @@
+export { default as i18nConfig } from "@/i18n/i18nConfig";
+export { default as initTranslations } from "@/i18n/initTranslations";
+export { default as TranslationsProvider } from "@/i18n/TranslationsProvider";
