+++ src/i18n/initTranslations.js
@@ -2,7 +2,7 @@ import { createInstance } from "i18next";
 import resourcesToBackend from "i18next-resources-to-backend";
 import { initReactI18next } from "react-i18next/initReactI18next";
 
-import i18nConfig from "@/i18nConfig";
+import { i18nConfig } from "@/i18n";
 
 export default async function initTranslations(
   locale,
