+++ src/middleware.js
@@ -1,6 +1,6 @@
 import { i18nRouter } from "next-i18n-router";
 
-import i18nConfig from "@/i18nConfig";
+import { i18nConfig } from "@/i18n";
 
 export function middleware(request) {
   return i18nRouter(request, i18nConfig);
