+++ tsconfig.json
@@ -28,9 +28,9 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     "src/middleware.js",
-    "src/app/i18n.js",
-    "src/components/common/TranslationsProvider.js",
-    "src/i18nConfig.js",
+    "src/i18n/initTranslations.js",
+    "src/i18n/TranslationsProvider.js",
+    "src/i18n/i18nConfig.js",
     "jest.config.js",
     "next.config.js",
     "commitlint.config.js",
